ファイル処理関数切り出しておく
